requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
ccxt==4.0.72
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
pandas==2.0.3
pycares==4.3.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
ta==0.10.2
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.6.2
yarl==1.9.2
